requirements.txt
absl-py==2.3.1
addict==2.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
argparse==1.4.0
asttokens==3.0.0
attrs==25.4.0
Automat==25.4.16
backcall==0.2.0
beautifulsoup4==4.14.2
buildtools==1.0.6
causal-conv1d==1.0.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==2.1.1
cityscapesscripts==2.2.4
click==8.3.0
clip==1.0
cloudpickle==3.1.2
colorama==0.4.6
coloredlogs==15.0.1
constantly==23.10.4
contourpy==1.3.3
crcmod==1.7
cryptography==46.0.3
cupy-cuda11x==13.6.0
cycler==0.12.1
dassl==0.6.3
decorator==5.2.1
dinov3==0.0.1
docopt==0.6.2
einops==0.8.1
entrypoints==0.3
executing==2.2.1
fastrlock==0.8.3
filelock==3.14.0
flake8==3.7.9
fonttools==4.60.1
fsspec==2025.9.0
ftfy==6.3.1
furl==2.1.4
future==1.0.0
fvcore==0.1.5.post20221221
gdown==5.2.0
greenlet==3.2.4
grpcio==1.76.0
huggingface-hub==0.21.4
humanfriendly==10.0
hyperlink==21.0.0
idna==3.4
incremental==24.7.2
iopath==0.1.10
ipython==8.12.3
isort==4.3.21
jedi==0.19.2
Jinja2==3.1.6
jmespath==0.10.0
joblib==1.5.2
kiwisolver==1.4.9
lightning-utilities==0.15.2
littleutils==0.2.4
loralib==0.1.2
mamba-ssm==2.2.4
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.7.3
matplotlib-inline==0.2.1
mccabe==0.6.1
mdurl==0.1.2
mmcv==2.2.0
mmcv-full==1.7.2
mmdet==3.3.0
mmengine==0.10.7
model-index==0.1.11
mpi4py==3.1.4
mpmath==1.3.0
networkx==3.5
ninja==1.13.0
numpy==1.26.4
ogb==1.3.6
omegaconf==2.3.0
open-clip-torch==3.2.0
opencv-python==4.8.0.76
opencv-python-headless==4.8.1.78
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.3
ordered-set==4.1.0
orderedmultidict==1.0.1
oss2==2.17.0
outdated==0.2.2
packaging==24.2
pandas==2.3.3
panopticapi==0.1
parso==0.8.5
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.3.0
pip==25.3
portalocker==3.2.0
prompt-toolkit==3.0.52
protobuf==6.33.0
ptyprocess==0.7.0
pure-eval==0.2.3
pycocotools==2.0.7
pycodestyle==2.5.0
pycparser==2.23
pycryptodome==3.23.0
pyflakes==2.1.1
Pygments==2.19.2
pyparsing==3.2.5
pyquaternion==0.9.9
PySocks==1.7.1
pysodmetrics==1.4.2
python-dateutil==2.9.0.post0
pytorch-wavelets==1.3.0
pytz==2023.4
PyWavelets==1.9.0
PyYAML==6.0.2
redo==3.0.0
regex==2025.11.3
requests==2.28.2
rich==13.4.2
safetensors==0.5.3
scikit-learn==1.3.0
scipy==1.16.0
segment-anything==1.0
setuptools==80.10.2
shapely==2.1.2
simplejson==3.20.2
six==1.17.0
soupsieve==2.8
spikingjelly==0.0.0.0.14
SQLAlchemy==2.0.44
stack-data==0.6.3
submitit==1.5.3
sympy==1.14.0
tabulate==0.9.0
tb-nightly==2.21.0a20251023
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
termcolor==3.2.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
timm==1.0.22
tokenizers==0.13.3
torch==2.1.1+cu118
torch-dct==0.1.6
torchaudio==2.1.1+cu118
torchmetrics==1.8.2
torchvision==0.16.1+cu118
tqdm==4.67.1
traitlets==5.14.3
transformers==4.31.0
triton==2.1.0
Twisted==25.5.0
typing==3.7.4.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.13
wcwidth==0.2.14
Werkzeug==3.1.3
wheel==0.46.3
wilds==1.2.2
yacs==0.1.8
yapf==0.29.0
zope.interface==8.1